Gold/requirements.txt
# Core Dependencies
watchdog>=4.0.0
schedule>=1.2.0
requests>=2.31.0

# Google/Gmail
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Browser Automation
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0